requirements.txt
opencv-python>=4.8.0
mediapipe>=0.10.13
google-generativeai>=0.3.0
numpy>=1.24.0
ultralytics>=8.0.0
python-dotenv>=1.0.0
